Re-adding r985 and r986.

+++ branches/branch_filesharing/src/ditool/Partition.h
@@ -24,14 +24,14 @@ class DiskImage;
 class Partition {
 public:
     int                      partNo;
-    int                      partIdx;
+    size_t                   partIdx;
     DiskImage*               im;
     const struct disk_label* dl;
     struct disk_partition&   part;
     uint32_t                 fsOff;
     
     Partition(void);
-    Partition(int partNo, int partIdx, DiskImage* im, const struct disk_label* dl, struct disk_partition& part);
+    Partition(int partNo, size_t partIdx, DiskImage* im, const struct disk_label* dl, struct disk_partition& part);
     ~Partition(void);
     
     int  readSectors(uint32_t sector, uint32_t count, uint8_t* dst) const;
